move test_socket_error_on_read
